+++ vscode-extension/test/e2e/defaults-sass/find-definition.test.js
@@ -1,6 +1,6 @@
+const assert = require("assert");
 const vscode = require("vscode");
 const { getDocUri, showFile, sleepCI } = require("./util");
-const { testDefinition } = require("./definition-helper");
 
 const styles = getDocUri("styles.sass");
 const tokens = getDocUri("core/_tokens.sass");
@@ -37,3 +37,36 @@ test("finds definition of a sass mixin included with at-rule", async () => {
 
 	await testDefinition(styles, new vscode.Position(7, 13), expected);
 });
+
+
+/**
+ * @param {import('vscode').Uri} docUri
+ * @param {import('vscode').Position} position
+ * @param {import('vscode').Location} expectedLocation
+ * @returns {Promise<void>}
+ */
+async function testDefinition(docUri, position, expectedLocation) {
+	await showFile(docUri);
+
+	const result =
+		/** @type {import('vscode').Location[]} */
+		(
+			await vscode.commands.executeCommand(
+				"vscode.executeDefinitionProvider",
+				docUri,
+				position,
+			)
+		);
+
+	if (result[0] === undefined) {
+		assert.fail("The 'result[0]' is undefined.");
+	}
+
+	assert.ok(
+		result[0].range.isEqual(expectedLocation.range),
+		`Expected ${JSON.stringify(result[0].range)} to equal ${JSON.stringify(
+			expectedLocation.range,
+		)} in ${docUri.fsPath}`,
+	);
+	assert.strictEqual(result[0].uri.fsPath, expectedLocation.uri.fsPath);
+}
